+++ reporter/commands.js
@@ -75,13 +75,13 @@ Cypress.Commands.add(
     }
 );
 Cypress.Commands.add('issue', { prevSubject: true }, (allure, name, url) => {
-    const baseUrl = Cypress.env('allureIssueUrl');
-    allure.issue(name, baseUrl ? `${baseUrl}/${url}` : url);
+    const issuePrefix = Cypress.env('issuePrefix');
+    allure.issue(name, issuePrefix ? `${issuePrefix}${url}` : url);
     cy.wrap(allure, { log: false });
 });
 Cypress.Commands.add('tms', { prevSubject: true }, (allure, name, url) => {
-    const baseUrl = Cypress.env('allureTmsUrl');
-    allure.tms(name, baseUrl ? `${baseUrl}/${url}` : url);
+    const tmsPrefix = Cypress.env('tmsPrefix');
+    allure.tms(name, tmsPrefix ? `${tmsPrefix}${url}` : url);
     cy.wrap(allure, { log: false });
 });
 Cypress.Commands.add(